+++ palace/utils/configfile.hpp
@@ -90,6 +90,16 @@ struct ElementData
 
 // Problem & Model Config.
 
+struct OutputFormatsData
+{
+public:
+  // Enable Paraview output format.
+  bool paraview = true;
+
+  // Enable MFEM GLVis grid function output format.
+  bool gridfunction = false;
+};
+
 struct ProblemData
 {
 public:
@@ -101,6 +111,9 @@ public:
 
   // Output path for storing results.
   std::string output = "";
+
+  // Output formats configuration.
+  OutputFormatsData output_formats = {};
 
   void SetUp(json &config);
 };
@@ -656,16 +669,6 @@ public:
 
 // Solver Configuration.
 
-struct OutputFormatsData
-{
-public:
-  // Enable Paraview output format.
-  bool paraview = true;
-
-  // Enable MFEM Grid Function output format.
-  bool mfem_grid_function = false;
-};
-
 struct DrivenSolverData
 {
 public:
@@ -690,9 +693,6 @@ public:
 
   // Memory required for adaptive sampling convergence.
   int adaptive_memory = 2;
-
-  // Output formats configuration.
-  OutputFormatsData output_formats = {};
 
   void SetUp(json &solver);
 };
@@ -761,9 +761,6 @@ public:
   // nonlinear eigenvalue solver.
   int max_restart = 2;
 
-  // Output formats configuration.
-  OutputFormatsData output_formats = {};
-
   void SetUp(json &solver);
 };
 
@@ -773,9 +770,6 @@ public:
   // Number of fields to write to disk.
   int n_post = 0;
 
-  // Output formats configuration.
-  OutputFormatsData output_formats = {};
-
   void SetUp(json &solver);
 };
 
@@ -784,9 +778,6 @@ struct MagnetostaticSolverData
 public:
   // Number of fields to write to disk.
   int n_post = 0;
-
-  // Output formats configuration.
-  OutputFormatsData output_formats = {};
 
   void SetUp(json &solver);
 };
@@ -821,9 +812,6 @@ public:
   // Adaptive time-stepping tolerances for CVODE and ARKODE.
   double rel_tol = 1e-4;
   double abs_tol = 1e-9;
-
-  // Output formats configuration.
-  OutputFormatsData output_formats = {};
 
   void SetUp(json &solver);
 };